Consolidate browser specs and poll for Reveal.js readiness
Merge 6 index.html browser spec files into one (presentation,
slide-zoom, chord-visibility, theme-toggle, multiplex-ui, TOC-nav)
to cut down on Capybara session setup/teardown overhead, and rename
print_spec.rb to print_html_spec.rb to match. Replace fixed sleep 1.5
calls with a wait_for_reveal poller that proceeds as soon as
Reveal.isReady() is true.
